requirements.txt
python-telegram-bot==22.6
python-dotenv==1.2.2
requests==2.32.5
